+++ CHANGELOG.md
@@ -1,11 +1,15 @@
 # CHANGELOG
 
 ## Unreleased
-- **Other Features**
-  - [spiral/scaffolder] Added new public method `declarationDirectory` to the `Spiral\Scaffolder\Config\ScaffolderConfig` 
-    class that returns the directory path of the specified declaration, or default directory path if not specified.
 - **Medium Impact Changes**
   - [spiral/scaffolder] Method `baseDirectory` of `Spiral\Scaffolder\Config\ScaffolderConfig` class is deprecated.
+- **Other Features**
+  - Added `Spiral\Auth\TokenStorageScope`, this class can be used to get the concrete implementation of
+    the token storage in a current container scope.
+  - [spiral/auth-http] Added a `Spiral\Auth\TokenStorageInterface` binding in the `Spiral\Auth\Middleware\AuthMiddleware` 
+    with the used TokenStorage.
+  - [spiral/scaffolder] Added new public method `declarationDirectory` to the `Spiral\Scaffolder\Config\ScaffolderConfig` 
+    class that returns the directory path of the specified declaration, or default directory path if not specified.
 
 ## 3.7.1 - 2023-04-21
 - **Bug Fixes**